refactor(units): SI length header (#4)
